Changed pattern generation to use setNote and addNote. Added length variable to pattern. Changed random in pattern creation to MyRand.
