+++ src/sequencer/track.cpp
@@ -12,7 +12,8 @@ Track::Track(eTrackType_t track_type, byte midi_channel) {
 };
 
 int Track::addPattern(int length) {
-  Patterns.emplace_back(Pattern(length));
+  ePatternType_t pattern = (_trackType == TRACK_DRUMS) ? DRUM : SYNTH;
+  Patterns.emplace_back(Pattern(length, pattern));
   return (Patterns.size()-1);
 }
 
@@ -25,6 +26,7 @@ Pattern* Track::getPattern(int patternIndex) {
 
 bool Track::addStackNote(int note, int length)
 {
+  DEBF("Add to stack note: %d, length: %d\r\n", note, length);
   for(unsigned char i = 0; i < NOTE_STACK_SIZE; i++) {
     if (_noteStack[i].length == -1) {
       _noteStack[i].length = length;
